+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/referee.js
@@ -86,6 +86,10 @@ if (!gwoRefereeChangesLoaded) {
           var aiBrain = aiInUse();
 
           var findAIPath = function (type) {
+            if (type === "cluster") {
+              return "/pa/ai_cluster/";
+            }
+
             var game = model.game();
             var ai = game.galaxy().stars()[game.currentStar()].ai();
             var inventory = game.inventory();
@@ -387,11 +391,26 @@ if (!gwoRefereeChangesLoaded) {
                   gwoUnit.skitterWeapon
                 );
 
+                var clusterArmyIndex = function () {
+                  if (ai.faction === 4) {
+                    return 0;
+                  } else if (ai.foes) {
+                    var index = _.findIndex(ai.foes, { faction: 4 });
+                    if (index !== -1) {
+                      return index + 1;
+                    }
+                  }
+                };
+
                 var inventory = self.game().inventory();
 
                 var units = parse(unitsGet[0]).units;
                 var aiUnitMap = parse(aiMapGet[0]);
                 var aiX1UnitMap = parse(aiX1MapGet[0]);
+                var clusterUnitMapPath =
+                  "/pa/ai_cluster/unit_maps/ai_unit_map.json";
+                var clusterUnitMapTitansPath =
+                  "/pa/ai_cluster/unit_maps/ai_unit_map_x1.json";
                 _.times(aiFactionCount, function (n) {
                   var currentCount = n;
                   var enemyAIUnitMap = GW.specs.genAIUnitMap(
@@ -407,10 +426,17 @@ if (!gwoRefereeChangesLoaded) {
                   GW.specs
                     .genUnitSpecs(aiSpecs, aiTag[n])
                     .then(function (aiSpecFiles) {
-                      var enemyAIUnitMapFile = aiUnitMapPath + aiTag[n];
+                      var unitMapPath = aiUnitMapPath;
+                      var unitMapTitansPath = aiUnitMapTitansPath;
+                      if (clusterArmyIndex() === currentCount) {
+                        unitMapPath = clusterUnitMapPath;
+                        unitMapTitansPath = clusterUnitMapTitansPath;
+                      }
+
+                      var enemyAIUnitMapFile = unitMapPath + aiTag[n];
                       var enemyAIUnitMapPair = {};
                       enemyAIUnitMapPair[enemyAIUnitMapFile] = enemyAIUnitMap;
-                      var enemyX1AIUnitMapFile = aiUnitMapTitansPath + aiTag[n];
+                      var enemyX1AIUnitMapFile = unitMapTitansPath + aiTag[n];
                       var enemyX1AIUnitMapPair = {};
                       enemyX1AIUnitMapPair[enemyX1AIUnitMapFile] =
                         enemyX1AIUnitMap;
@@ -422,6 +448,7 @@ if (!gwoRefereeChangesLoaded) {
                         ? _.assign(enemyX1AIUnitMapPair, aiSpecFiles)
                         : {};
                       var aiFiles = _.assign({}, aiFilesClassic, aiFilesX1);
+
                       if (ai.inventory) {
                         var aiInventory = [];
                         aiInventory =
@@ -574,6 +601,21 @@ if (!gwoRefereeChangesLoaded) {
               });
             });
             return done.promise();
+          };
+
+          var isClusterPresent = function (inventory, ai) {
+            if (inventory.getTag("global", "playerFaction") === 4) {
+              return "Player";
+            } else if (ai.faction === 4 && ai.mirrorMode !== true) {
+              return "Enemy";
+            } else if (ai.foes) {
+              for (var foe of ai.foes) {
+                if (foe.faction === 4) {
+                  return "Enemy";
+                }
+              }
+            }
+            return "None";
           };
 
           // parse AI mods and load the results into self.files()
@@ -735,8 +777,28 @@ if (!gwoRefereeChangesLoaded) {
             var alliedCommanders = _.isUndefined(ai.ally)
               ? inventory.minions()
               : inventory.minions().concat(ai.ally);
+            var clusterCommanders = ["SupportPlatform", "SupportCommander"];
+            var clusterAIMods = _.map(clusterCommanders, function (commander) {
+              return {
+                type: "factory",
+                op: "replace",
+                toBuild: commander,
+                idToMod: "priority",
+                value: 0,
+              };
+            });
+            var clusterAIPath = findAIPath("cluster");
 
-            var parseFiles = function (aiPath, promise, aiToModify) {
+            var aiPathCreation = function (aiPath, filePath, cullLength) {
+              return aiPath + filePath.slice(cullLength);
+            };
+
+            var parseFiles = function (
+              aiPath,
+              promise,
+              aiToModify,
+              clusterPresent
+            ) {
               api.file.list(aiPath, true).then(function (fileList) {
                 var configFiles = self.files();
                 var queue = [];
@@ -768,8 +830,8 @@ if (!gwoRefereeChangesLoaded) {
 
                 _.forEach(fileList, function (filePath) {
                   if (
-                    !_.endsWith(filePath, ".json") &&
-                    _.includes(filePath, "/neural_networks/") &&
+                    !_.endsWith(filePath, ".json") ||
+                    _.includes(filePath, "/neural_networks/") ||
                     _.endsWith(filePath, "ai_config.json")
                   ) {
                     return;
@@ -778,6 +840,7 @@ if (!gwoRefereeChangesLoaded) {
                   var deferred2 = $.Deferred();
 
                   var aiBuildOps = [];
+                  var aiClusterOps = [];
                   var quellerSubCommander = false;
                   if (
                     isQueller &&
@@ -788,12 +851,12 @@ if (!gwoRefereeChangesLoaded) {
                     quellerSubCommander = true;
                   }
 
+                  // Only mods associated with the file's AI manager are loaded
                   if (
                     aiToModify !== "None" &&
                     !_.isEmpty(aiMods[1]) &&
                     (!isQueller || quellerSubCommander || aiToModify === "All")
                   ) {
-                    // Only mods associated with the file's AI manager are loaded
                     if (_.includes(filePath, "/fabber_builds/")) {
                       aiBuildOps = _.filter(aiMods[1], {
                         type: "fabber",
@@ -808,11 +871,29 @@ if (!gwoRefereeChangesLoaded) {
                       });
                     }
                   }
+                  if (_.includes(filePath, "/factory_builds/")) {
+                    if (clusterPresent === "Enemy") {
+                      aiClusterOps = clusterAIMods;
+                    } else if (clusterPresent === "Player") {
+                      aiBuildOps = aiBuildOps.concat(clusterAIMods);
+                    }
+                  }
 
                   queue.push(deferred2);
 
                   $.getJSON("coui:/" + filePath)
                     .then(function (json) {
+                      if (clusterPresent === "Enemy") {
+                        aiFilePath = aiPathCreation(
+                          clusterAIPath,
+                          filePath,
+                          findAIPath("enemy").length
+                        );
+                        var clusterJson = _.cloneDeep(json);
+                        addTechToAI(clusterJson, aiClusterOps);
+                        configFiles[aiFilePath] = clusterJson;
+                      }
+
                       if (aiToModify === "All") {
                         if (!_.isEmpty(aiBuildOps)) {
                           addTechToAI(json, aiBuildOps);
@@ -822,29 +903,38 @@ if (!gwoRefereeChangesLoaded) {
                           if (isQueller) {
                             // We don't know if the aiPath contains q_uber
                             var quellerEnemyPath = findAIPath("enemy");
-                            filePath =
-                              quellerEnemyPath +
-                              filePath.slice(aiTechPath.length);
-                            configFiles[filePath] = json;
+                            aiFilePath = aiPathCreation(
+                              quellerEnemyPath,
+                              filePath,
+                              aiTechPath.length
+                            );
+                            configFiles[aiFilePath] = json;
                             if (quellerSubCommander) {
-                              filePath =
-                                subcommanderAIPath +
-                                filePath.slice(quellerEnemyPath.length);
-                              configFiles[filePath] = json;
+                              aiFilePath = aiPathCreation(
+                                subcommanderAIPath,
+                                filePath,
+                                quellerEnemyPath.length
+                              );
+                              configFiles[aiFilePath] = json;
                             }
                           } else {
-                            filePath =
-                              aiPath + filePath.slice(aiTechPath.length);
-                            configFiles[filePath] = json;
+                            aiFilePath = aiPathCreation(
+                              aiPath,
+                              filePath,
+                              aiTechPath.length
+                            );
+                            configFiles[aiFilePath] = json;
                           }
                         } else {
                           configFiles[filePath] = json;
                         }
                       } else if (aiToModify === "SubCommanders") {
+                        aiFilePath = filePath;
                         // Set up enemy AI first
                         if (!_.startsWith(filePath, aiTechPath)) {
-                          configFiles[filePath] = _.cloneDeep(json);
+                          configFiles[aiFilePath] = _.cloneDeep(json);
                         }
+                        configFiles[aiFilePath] = _.cloneDeep(json);
                         // Set up Sub Commanders
                         if (!_.isEmpty(aiBuildOps)) {
                           addTechToAI(json, aiBuildOps);
@@ -852,18 +942,20 @@ if (!gwoRefereeChangesLoaded) {
                         if (quellerSubCommander) {
                           // Put "load" files where Queller expects them to be
                           if (_.startsWith(filePath, aiTechPath)) {
-                            filePath =
-                              subcommanderAIPath +
-                              filePath.slice(aiTechPath.length);
+                            aiFilePath = aiPathCreation(
+                              subcommanderAIPath,
+                              filePath,
+                              aiTechPath.length
+                            );
                           }
                         } else {
                           // Titans/Penchant Sub Commanders share an ai_path with the enemy so need a new one
                           if (_.startsWith(filePath, aiPath)) {
-                            filePath =
+                            aiFilePath =
                               aiTechPath + filePath.slice(aiPath.length);
                           }
                         }
-                        configFiles[filePath] = json;
+                        configFiles[aiFilePath] = json;
                       } else {
                         configFiles[filePath] = json;
                       }
@@ -884,15 +976,21 @@ if (!gwoRefereeChangesLoaded) {
               alliedCommanders.length > 0
                 ? findAIPath("all")
                 : findAIPath("enemy");
+            var clusterPresence = isClusterPresent(inventory, ai);
 
-            if (_.isEmpty(inventory.aiMods())) {
-              parseFiles(aiFilePath, deferredAIFiles, "None");
+            if (_.isEmpty(inventory.aiMods() && clusterPresence !== "Player")) {
+              parseFiles(aiFilePath, deferredAIFiles, "None", clusterPresence);
             } else if (ai.mirrorMode === true) {
               parseFiles(aiFilePath, deferredAIFiles, "All");
             } else if (alliedCommanders.length > 0) {
-              parseFiles(aiFilePath, deferredAIFiles, "SubCommanders");
+              parseFiles(
+                aiFilePath,
+                deferredAIFiles,
+                "SubCommanders",
+                clusterPresence
+              );
             } else {
-              parseFiles(aiFilePath, deferredAIFiles, "None");
+              parseFiles(aiFilePath, deferredAIFiles, "None", clusterPresence);
             }
 
             $.when(deferredAIFiles).then(function () {
@@ -1020,9 +1118,15 @@ if (!gwoRefereeChangesLoaded) {
             adjustAdvEcoMod(ai, aiBrain);
 
             var enemyAIPath = findAIPath("enemy");
+            var clusterAIPath = findAIPath("cluster");
+            var aiIsCluster = ai.faction === 4;
 
             // Avoid breaking enemies from earlier versions
-            ai.personality.ai_path = enemyAIPath;
+            if (aiIsCluster) {
+              ai.personality.ai_path = clusterAIPath;
+            } else {
+              ai.personality.ai_path = enemyAIPath;
+            }
 
             var slotsArrayAI = [];
             _.times(
@@ -1047,7 +1151,11 @@ if (!gwoRefereeChangesLoaded) {
               adjustAdvEcoMod(minion, aiBrain);
 
               // Avoid breaking enemies from earlier versions
-              minion.personality.ai_path = enemyAIPath;
+              if (aiIsCluster) {
+                minion.personality.ai_path = clusterAIPath;
+              } else {
+                minion.personality.ai_path = enemyAIPath;
+              }
 
               var slotsArrayMinions = [];
               _.times(
@@ -1076,7 +1184,12 @@ if (!gwoRefereeChangesLoaded) {
               adjustAdvEcoMod(foe, aiBrain);
 
               // Avoid breaking enemies from earlier versions
-              foe.personality.ai_path = enemyAIPath;
+              var foeIsCluster = foe.faction === 4;
+              if (foeIsCluster) {
+                foe.personality.ai_path = clusterAIPath;
+              } else {
+                foe.personality.ai_path = enemyAIPath;
+              }
 
               var slotsArrayFoes = [];
               _.times(