+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Cluster systems having more than one Worker army
+- Cluster AIs attempting to build Angels and Colonels
 
 ## v5.39.0 - 2022-07-04